clean up __init__ files; move PhaseEstimationCircuit under circuits/
