+++ qiskit/aqua/algorithms/single_sample/grover/grover.py
@@ -29,7 +29,7 @@ from qiskit.aqua import AquaError, Pluggable, PluggableType, get_pluggable_class
 from qiskit.aqua.utils import get_subsystem_density_matrix
 from qiskit.aqua.algorithms import QuantumAlgorithm
 from qiskit.aqua.components.initial_states import Custom
-
+from qiskit.aqua.circuits.gates import mct
 
 logger = logging.getLogger(__name__)
