+++ qiskit/aqua/circuits/gates/boolean_logic_gates.py
@@ -25,6 +25,7 @@ from qiskit import QuantumCircuit, QuantumRegister
 from qiskit.qasm import pi
 
 from qiskit.aqua import AquaError
+from .mct import mct
 
 logger = logging.getLogger(__name__)
