+++ qiskit/aqua/components/reciprocals/long_division.py
@@ -15,9 +15,12 @@
 # limitations under the License.
 # =============================================================================
 
-from qiskit import QuantumRegister, QuantumCircuit
-from qiskit.aqua.components.reciprocals import Reciprocal
 import numpy as np
+
+from qiskit import QuantumRegister, QuantumCircuit
+
+from qiskit.aqua.components.reciprocals import Reciprocal
+from qiskit.aqua.circuits.gates import mct
 
 
 class LongDivision(Reciprocal):
