+++ qiskit/aqua/components/reciprocals/lookup_rotation.py
@@ -20,8 +20,11 @@
 import itertools
 import logging
 import numpy as np
+
 from qiskit import QuantumRegister, QuantumCircuit
+
 from qiskit.aqua.components.reciprocals import Reciprocal
+from qiskit.aqua.circuits.gates import mct
 
 logger = logging.getLogger(__name__)
